Pass embedding request to the shortcode rendering controller
Traditionally, you have self-contained shortcodes that carry all information
necessary to render them as parameters. But sometimes, the information is in
the context of the embedding entity. E.g. if you want to embed a list of the
child nodes of the current node in a tree, it's more comfortable to use a short
code like [children] than [children id-of-current-node].

By passing the embedding request to the shortcode rendering controllers, we
allow them to become context sensitive.