Prepare transition from XML- to PHP-based routing configuration (#43)
XML-based routing configuration has been deprecated in Symfony 7.4.

To provide a way forward, this PR adds a PHP-based routing
configuration. The XML file will include the PHP version, and a
deprecation notice will be triggered to remind users that they need to
update their configuration.

The XML file will be removed in the future.